Set default target name for sonarqube rule
From this commit the sonarqube rule target name attribute  is optional
It is needed only when multiple `sonarqube` rule is used in the same `BUILD` package.

If not specified `sonarscan` target name is used.

Additionally removed `visibility = ["//visibility:public"],`
from README.md, as it not mandatory, and it is safer
(it is not allowing to use filegroup from other module)